requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
defusedxml==0.7.1
Django==3.2.19
django-cors-headers==4.0.0
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
importlib-metadata==6.6.0
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
Pillow==9.5.0
Pygments==2.15.1
PySocks==1.7.1
pytz==2023.3
requests==2.30.0
requests-toolbelt==1.0.0
rich==13.3.5
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.2
zipp==3.15.0
